refactor: CRTableList --> CRTableListItem; move: TableListPage
