+++ static_report/src/components/shared/Tables/TableList/CRTableListItem/CRTableListAssertions.tsx
