+++ static_report/src/components/shared/Tables/TableList/CRTableListItem/CRTableListColumnsSummary.tsx
